WIP: Refactor animations based on sync with Jimmy
- add react-transition-group
- upgrade @urbit/roller-api
- extract `ActivateCodeForm`
