+++ package.json
@@ -70,6 +70,7 @@
     "react-final-form-arrays": "^3.1.0",
     "react-scripts": "4.0.3",
     "react-teleporter": "^2.2.0",
+    "react-transition-group": "^4.4.2",
     "secp256k1": "^4.0.2",
     "styled-components": "^5.3.0",
     "styled-system": "^5.1.5",
@@ -87,6 +88,7 @@
     "@testing-library/react": "^12.0.0",
     "@types/keccak": "^3.0.1",
     "@types/lodash": "^4.14.170",
+    "@types/react-transition-group": "^4.4.2",
     "eslint-config-prettier": "^5.0.0",
     "eslint-plugin-prettier": "^3.1.0",
     "eslint-plugin-react-hooks": "^1.6.0",
