+++ src/style/index.scss
@@ -12,3 +12,6 @@
 
 /* Bridge Styles */
 @import './bridge.scss';
+
+/* Activate Flow animations */
+@import './activate.scss';
